feat! : added DI container, storage, etc. FULLY refactored project structure to CLEAN ARCHITECTURE, started refactoring of auth
