+++ .gitignore
@@ -22,5 +22,5 @@ generated/
 # Игнорировать временные файлы операционной системы
 .DS_Store
 
-
+*.hprof
 .env
