+++ app/src/main/java/com/example/ep2024/data/repository/UserRepositoryImpl.java
@@ -1,0 +1,14 @@
+package com.example.ep2024.data.repository;
+
+import com.example.ep2024.domain.model.Role;
+import com.example.ep2024.domain.model.user.User;
+import com.example.ep2024.domain.repository.UserRepository;
+
+public class UserRepositoryImpl implements UserRepository {
+
+    @Override
+    public User getUser(String id) {
+        return new User(1, "1234", Role.ADMIN);
+    }
+
+}
