+++ app/src/main/java/com/example/ep2024/domain/model/Category.kt
@@ -1,0 +1,14 @@
+package com.example.ep2024.domain.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+import java.util.UUID
+
+//TODO: not final
+@Parcelize
+data class Category(
+    val id: String = UUID.randomUUID().toString(),
+    val name: String,
+    val description: String,
+    val isActive: Boolean = true
+) : Parcelable
