+++ app/src/main/java/com/example/ep2024/domain/repository/ServiceRepository.java
@@ -1,0 +1,8 @@
+package com.example.ep2024.domain.repository;
+
+import com.example.ep2024.domain.model.Service;
+
+public interface ServiceRepository {
+    Service getService(String id);
+
+}
