+++ app/src/main/java/com/example/ep2024/domain/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.example.ep2024.domain.repository;
+
+import com.example.ep2024.domain.model.user.User;
+
+public interface UserRepository {
+    User getUser(String id);
+
+}
