+++ app/src/main/java/com/example/ep2024/presentation/adapter/EventAdapter.kt
@@ -1,0 +1,50 @@
+package com.example.ep2024.presentation.adapter
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.example.ep2024.databinding.ItemEventBinding
+import com.example.ep2024.domain.model.Event
+
+class EventAdapter(private var events: List<Event>) :
+    RecyclerView.Adapter<EventAdapter.EventViewHolder>() {
+
+    private var onItemClickListener: ((Event) -> Unit)? = null
+
+    fun setOnItemClickListener(listener: (Event) -> Unit) {
+        onItemClickListener = listener
+    }
+
+    fun updateList(newList: List<Event>) {
+        events = newList
+        notifyDataSetChanged()
+    }
+
+    inner class EventViewHolder(private val binding: ItemEventBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+
+        fun bind(event: Event) {
+            binding.textViewEventName.text = event.name
+            binding.textViewEventDescription.text = event.description
+            //TODO:
+            binding.root.setOnClickListener {
+                onItemClickListener?.invoke(event)
+            }
+        }
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): EventViewHolder {
+        val binding = ItemEventBinding.inflate(
+            LayoutInflater.from(parent.context),
+            parent,
+            false
+        )
+        return EventViewHolder(binding)
+    }
+
+    override fun onBindViewHolder(holder: EventViewHolder, position: Int) {
+        holder.bind(events[position])
+    }
+
+    override fun getItemCount(): Int = events.size
+}
