--- app/src/main/java/com/example/ep2024/utils/category/SuggestionStatus.kt
@@ -1,9 +1,0 @@
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-enum class SuggestionStatus : Parcelable {
-    PENDING,
-    APPROVED,
-    REJECTED
-}
