--- app/src/main/java/com/example/ep2024/utils/service/ConfirmationMethod.kt
@@ -1,10 +1,0 @@
-package com.example.ep2024.utils.service
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-enum class ConfirmationMethod : Parcelable {
-    AUTOMATIC,
-    MANUAL
-}
