+++ gradle/libs.versions.toml
@@ -2,6 +2,8 @@
 agp = "8.9.0"
 androidxNavigationFragmentKtx = "2.5.3"
 cardview = "1.0.0"
+converterGson = "2.9.0"
+dagger = "2.40.5"
 kotlin = "2.0.0"
 coreKtx = "1.15.0"
 junit = "4.13.2"
@@ -12,12 +14,16 @@ lifecycleRuntimeKtx = "2.8.7"
 activityCompose = "1.9.3"
 composeBom = "2024.04.01"
 fragment = "1.8.5"
+loggingInterceptor = "4.9.3"
 material = "1.12.0"
 navigationFragmentKtx = "2.8.4"
 navigationFragmentKtxVersion = "2.7.4"
 navigationUiKtx = "2.7.4"
 navigationUiKtxVersion = "2.5.3"
 recyclerview = "1.3.2"
+retrofit = "2.9.0"
+securityCrypto = "1.1.0-alpha06"
+securityCryptoVersion = "1.0.0"
 
 [libraries]
 androidx-cardview = { module = "androidx.cardview:cardview", version.ref = "cardview" }
@@ -29,6 +35,10 @@ androidx-navigation-ui-ktx = { module = "androidx.navigation:navigation-ui-ktx",
 androidx-navigation-ui-ktx-v253 = { module = "androidx.navigation:navigation-ui-ktx", version.ref = "navigationUiKtxVersion" }
 androidx-navigation-ui-ktx-v274 = { module = "androidx.navigation:navigation-ui-ktx", version.ref = "navigationUiKtx" }
 androidx-recyclerview = { module = "androidx.recyclerview:recyclerview", version.ref = "recyclerview" }
+androidx-security-crypto = { module = "androidx.security:security-crypto", version.ref = "securityCrypto" }
+converter-gson = { module = "com.squareup.retrofit2:converter-gson", version.ref = "converterGson" }
+dagger = { module = "com.google.dagger:dagger", version.ref = "dagger" }
+dagger-compiler = { module = "com.google.dagger:dagger-compiler", version.ref = "dagger" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
 androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
 androidx-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espressoCore" }
@@ -44,7 +54,10 @@ androidx-ui-test-junit4 = { group = "androidx.compose.ui", name = "ui-test-junit
 androidx-material3 = { group = "androidx.compose.material3", name = "material3" }
 androidx-fragment = { group = "androidx.fragment", name = "fragment", version.ref = "fragment" }
 kotlin-parcelize-runtime = { module = "org.jetbrains.kotlin:kotlin-parcelize-runtime", version.ref = "kotlinParcelizeRuntime" }
+logging-interceptor = { module = "com.squareup.okhttp3:logging-interceptor", version.ref = "loggingInterceptor" }
 material = { module = "com.google.android.material:material", version.ref = "material" }
+retrofit = { module = "com.squareup.retrofit2:retrofit", version.ref = "retrofit" }
+security-crypto = { group = "androidx.security", name = "security-crypto", version.ref = "securityCryptoVersion" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
